backend now receiving id

+++ src/main/java/com/example/accountDetails/utils/AccountDetailsRepository.java
@@ -1,6 +1,5 @@
 package com.example.accountDetails.utils;
 
-
 import com.example.accountDetails.pojo.AccountDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,11 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AccountDetailsRepository extends JpaRepository<AccountDetails, Long> {
+public interface AccountDetailsRepository extends JpaRepository<AccountDetails, Integer> {
 
-    // Find Account Details by first name and last name
+    // Custom query to find an account by first name and last name
     Optional<AccountDetails> findByFirstNameAndLastName(String firstName, String lastName);
 
-    // You don't need custom save, delete, and findAll methods
-    // as JpaRepository already provides these operations.
+    // Additional custom queries can be added as needed
 }
